% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{GST.perm}
\alias{GST.perm}
\title{GST.perm Function}
\usage{
GST.perm(dst, sizes, Q, alpha, B)
}
\arguments{
\item{dst}{is the distance matrix of the data set}

\item{sizes}{is a vector of the sample sizes}

\item{Q}{is the distance threshold parameter}

\item{alpha}{is the significance level}

\item{B}{is the number of replicates in the permutation test}
}
\value{
\code{TRUE} if it decides to reject the EBT null hypothesis or \code{FALSE} if not
}
\description{
Runs a permutation test using the Energy Distance test statistic
}
\examples{
x = matrix(rexp(100 * 2), 100, 2)
y = matrix(rexp(200 * 2), 200, 2)
z <- rbind(x,y)
dst <- as.matrix(dist(z))
decision <- GST.perm(dst, c(100, 200),2, 0.05, 1000)

}
