% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{GST}
\alias{GST}
\title{GST Function}
\usage{
GST(dst, ix, sizes, Q)
}
\arguments{
\item{dst}{is the distance matrix of the data set}

\item{ix}{is a permutation of row indices of z}

\item{sizes}{is a vector of the sample sizes}

\item{Q}{is the distance threshold parameter}
}
\value{
The GST test statistic
}
\description{
Calculates the test statistic for the two-sampled Graph Based Test
}
\examples{
x = matrix(rexp(100 * 2), 100, 2)
y = matrix(rexp(200 * 2), 200, 2)
z <- rbind(x,y)
dst <- as.matrix(dist(z))
GST_statistic <- GST(dst, 1:nrow(dst), c(100, 200), 2)
}
