% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{HTT.perm}
\alias{HTT.perm}
\title{HTT.perm Function}
\usage{
HTT.perm(z, sizes, alpha, B)
}
\arguments{
\item{z}{is the matrix of the data set (each row is an observation)}

\item{sizes}{is a vector of the sample sizes}

\item{alpha}{is the significance level}

\item{B}{is the number of replicates in the permutation test}
}
\value{
\code{TRUE} if it decides to reject the HTT null hypothesis or \code{FALSE} if not
}
\description{
Runs a permutation test using the Hotelling's T-squared statistic
}
\examples{
x = matrix(rexp(100 * 2), 100, 2)
y = matrix(rexp(200 * 2), 200, 2)
z <- rbind(x,y)
decision <- HTT.perm(z, c(100, 200), 0.05, 1000)
}
