% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{HTT}
\alias{HTT}
\title{HTT Function}
\usage{
HTT(z, ix, sizes)
}
\arguments{
\item{z}{is the matrix of the data set (each row is an observation)}

\item{ix}{is a permutation of row indices of z}

\item{sizes}{is a vector of the sample sizes}
}
\value{
The HTT test statistic
}
\description{
Calculates the test statistic for the two-sampled Hotelling's T-squared Test
}
\examples{
x = matrix(rexp(100 * 2), 100, 2)
y = matrix(rexp(200 * 2), 200, 2)
z <- rbind(x,y)
HTT_statistic <- HTT(z, 1:nrow(z), c(100, 200))
}
