% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{NNT.perm}
\alias{NNT.perm}
\title{NNT.perm Function}
\usage{
NNT.perm(z, sizes, R, alpha, B)
}
\arguments{
\item{z}{is the matrix of the data set (each row is an observation)}

\item{sizes}{is a vector of the sample sizes}

\item{R}{is the number of neighbors to look at}

\item{alpha}{is the significance level}

\item{B}{is the number of replicates in the permutation test}
}
\value{
\code{TRUE} if it decides to reject the NNT null hypothesis or \code{FALSE} if not
}
\description{
Runs a permutation test using the Nearest Neighbors test statistic
}
\examples{
x = matrix(rexp(100 * 2), 100, 2)
y = matrix(rexp(200 * 2), 200, 2)
z <- rbind(x,y)
decision <- NNT.perm(z, c(100, 200), 4, 0.05, 1000)
}
