% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{NNT}
\alias{NNT}
\title{NNT Function}
\usage{
NNT(z, ix, sizes, R)
}
\arguments{
\item{z}{is the matrix of the data set (each row is an observation)}

\item{ix}{is a permutation of row indices of z}

\item{sizes}{is a vector of the sample sizes}

\item{R}{is the number of neighbors to look at}
}
\value{
The NNT test statistic
}
\description{
Calculates the test statistic for the two-sampled Nearest Neighbor Test
}
\examples{
x = matrix(rexp(100 * 2), 100, 2)
y = matrix(rexp(200 * 2), 200, 2)
z <- rbind(x,y)
NNT_statistic <- NNT(z, 1:nrow(z), c(100, 200), 4)
}
